#152 - refactoring, restructured package query.crs
