+++ src/main/java/org/opengis/cite/ogcapifeatures10/conformance/crs/query/crs/feature/FeatureValidCrsParameter.java
@@ -1,4 +1,4 @@
-package org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs;
+package org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.feature;
 
 import static io.restassured.http.Method.GET;
 import static org.opengis.cite.ogcapifeatures10.EtsAssert.assertCrsHeader;
@@ -6,17 +6,7 @@ import static org.opengis.cite.ogcapifeatures10.OgcApiFeatures10.CRS_PARAMETER;
 import static org.opengis.cite.ogcapifeatures10.OgcApiFeatures10.GEOJSON_MIME_TYPE;
 import static org.opengis.cite.ogcapifeatures10.util.JsonUtils.findFeatureUrlForGeoJson;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-
-import org.opengis.cite.ogcapifeatures10.conformance.CommonFixture;
-import org.opengis.cite.ogcapifeatures10.conformance.SuiteAttribute;
-import org.opengis.cite.ogcapifeatures10.util.JsonUtils;
-import org.testng.ITestContext;
 import org.testng.SkipException;
-import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import io.restassured.path.json.JsonPath;
@@ -46,25 +36,7 @@ import io.restassured.response.Response;
  *
  * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
  */
-public class FeatureValidCrsParameter extends CommonFixture {
-
-    @DataProvider(name = "collectionFeatureId")
-    public Iterator<Object[]> collectionFeatureId( ITestContext testContext ) {
-        Map<String, JsonPath> collectionsResponses = (Map<String, JsonPath>) testContext.getSuite().getAttribute( SuiteAttribute.COLLECTION_TO_ID.getName() );
-        Map<String, String> collectionNameToFeatureId = (Map<String, String>) testContext.getSuite().getAttribute( SuiteAttribute.FEATUREIDS.getName() );
-        List<Object[]> collectionsData = new ArrayList<>();
-        for ( Map.Entry<String, JsonPath> collection : collectionsResponses.entrySet() ) {
-            String collectionId = collection.getKey();
-            String featureId = null;
-            if ( collectionNameToFeatureId != null )
-                featureId = collectionNameToFeatureId.get( collectionId );
-            JsonPath json = collection.getValue();
-            for ( String crs : JsonUtils.parseAsList( "crs", json ) ) {
-                collectionsData.add( new Object[] { collectionId, json, featureId, crs } );
-            }
-        }
-        return collectionsData.iterator();
-    }
+public class FeatureValidCrsParameter extends AbstractFeatureCrs {
 
     /**
      * Test: Content-Crs header in the path /collections/{collectionId}/items/{featureId}
@@ -79,7 +51,7 @@ public class FeatureValidCrsParameter extends CommonFixture {
      *            the crs to test, never <code>null</code>
      */
     @Test(description = "Implements A.2.1 Query, Parameter crs, Abstract Test 1 (Requirement /req/crs/fc-crs-definition, /req/crs/fc-crs-valid-value B, /req/crs/ogc-crs-header, /req/crs/ogc-crs-header-value, /req/crs/geojson), "
-                        + "Content-Crs header in the path /collections/{collectionId}/items/{featureId}", dataProvider = "collectionFeatureId", dependsOnGroups = "crs-conformance")
+                        + "Content-Crs header in the path /collections/{collectionId}/items/{featureId}", dataProvider = "collectionFeatureIdCrs", dependsOnGroups = "crs-conformance")
     public void verifyFeaturePathCrsHeader( String collectionId, JsonPath collection, String featureId, String crs ) {
         String featureUrl = findFeatureUrlForGeoJson( rootUri, collection, featureId );
         if ( featureUrl == null )
