+++ src/main/java/org/opengis/cite/ogcapifeatures10/conformance/crs/query/crs/features/AbstractFeaturesCrs.java
@@ -1,0 +1,47 @@
+package org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.features;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+
+import org.opengis.cite.ogcapifeatures10.conformance.CommonFixture;
+import org.opengis.cite.ogcapifeatures10.conformance.SuiteAttribute;
+import org.opengis.cite.ogcapifeatures10.util.JsonUtils;
+import org.testng.ITestContext;
+import org.testng.annotations.DataProvider;
+
+import io.restassured.path.json.JsonPath;
+
+/**
+ * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
+ */
+public class AbstractFeaturesCrs extends CommonFixture {
+
+    @DataProvider(name = "collectionIdAndJson")
+    public Iterator<Object[]> collectionIdAndJson( ITestContext testContext ) {
+        Map<String, JsonPath> collectionsResponses = (Map<String, JsonPath>) testContext.getSuite().getAttribute( SuiteAttribute.COLLECTION_TO_ID.getName() );
+        List<Object[]> collectionsData = new ArrayList<>();
+        for ( Map.Entry<String, JsonPath> collection : collectionsResponses.entrySet() ) {
+            String collectionId = collection.getKey();
+            JsonPath json = collection.getValue();
+            collectionsData.add( new Object[] { collectionId, json } );
+        }
+        return collectionsData.iterator();
+    }
+
+    @DataProvider(name = "collectionIdAndJsonAndCrs")
+    public Iterator<Object[]> collectionIdAndJsonAndCrs( ITestContext testContext ) {
+        Map<String, JsonPath> collectionsResponses = (Map<String, JsonPath>) testContext.getSuite().getAttribute( SuiteAttribute.COLLECTION_TO_ID.getName() );
+        List<Object[]> collectionsData = new ArrayList<>();
+        for ( Map.Entry<String, JsonPath> collection : collectionsResponses.entrySet() ) {
+            String collectionId = collection.getKey();
+            JsonPath json = collection.getValue();
+            for ( String crs : JsonUtils.parseAsList( "crs", json ) ) {
+                collectionsData.add( new Object[] { collectionId, json, crs } );
+            }
+        }
+        return collectionsData.iterator();
+    }
+
+}
