+++ src/main/resources/org/opengis/cite/ogcapifeatures10/testng.xml
@@ -27,7 +27,8 @@
     <packages>
       <package name="org.opengis.cite.ogcapifeatures10.conformance.crs" />
       <package name="org.opengis.cite.ogcapifeatures10.conformance.crs.discovery" />
-      <package name="org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs" />
+      <package name="org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.features" />
+      <package name="org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.feature" />
     </packages>
   </test>
