+++ src/test/java/org/opengis/cite/ogcapifeatures10/conformance/crs/query/crs/FeatureInvalidCrsParameterTest.java
@@ -7,7 +7,7 @@ import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.endsWith;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.FeaturesInvalidCrsParameter.UNSUPPORTED_CRS;
+import static org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.features.FeaturesInvalidCrsParameter.UNSUPPORTED_CRS;
 
 import java.net.URLEncoder;
 
@@ -15,6 +15,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.feature.FeatureInvalidCrsParameter;
 import org.testng.ISuite;
 import org.testng.ITestContext;
