+++ src/test/java/org/opengis/cite/ogcapifeatures10/conformance/crs/query/crs/FeaturesValidCrsParameterTest.java
@@ -14,6 +14,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.features.FeaturesValidCrsParameter;
 import org.testng.ISuite;
 import org.testng.ITestContext;
